requirements.txt
attrs==19.3.0
beautifulsoup4==4.9.1
click==7.1.2
jdatetime==3.6.2
matplotlib==3.2.2
packaging==20.4
py==1.9.0
PyInstaller==3.6
pyparsing==2.4.7
pytest==5.4.3
python-cfonts==1.3.0
python-dateutil==2.8.1
requests==2.24.0
requests-futures==1.0.0
termtables==0.2.1
Unidecode==1.1.1